+++ code/migration/src/main/scala/com/namely/chiefofstate/migration/versions/v5/V5.scala
@@ -11,13 +11,13 @@ import com.namely.protobuf.chiefofstate.v1.persistence.{ EventWrapper, StateWrap
 import org.slf4j.{ Logger, LoggerFactory }
 import slick.basic.DatabaseConfig
 import slick.dbio.DBIO
-import slick.jdbc.{ GetResult, ResultSetConcurrency, ResultSetType, JdbcProfile }
+import slick.jdbc.{ GetResult, JdbcProfile, ResultSetConcurrency, ResultSetType }
 import slick.jdbc.PostgresProfile.api._
 import V5.log
 import scala.util.Try
 import akka.stream.scaladsl.Source
 import scala.concurrent.duration.Duration
-import scala.concurrent.{Future, Await}
+import scala.concurrent.{ Await, Future }
 import akka.actor.typed.ActorSystem
 import akka.Done
 import com.namely.protobuf.chiefofstate.plugins.persistedheaders.v1.headers.{
