+++ code/migration/src/test/scala/com/namely/chiefofstate/migration/versions/v2/V2Spec.scala
@@ -21,6 +21,7 @@ import org.testcontainers.utility.DockerImageName
 import slick.basic.DatabaseConfig
 import slick.jdbc.{ JdbcBackend, JdbcProfile }
 import slick.jdbc.PostgresProfile.api._
+import com.namely.chiefofstate.migration.helper.DbHelper._
 
 import java.sql.{ Connection, DriverManager }
 import scala.concurrent.{ Await, ExecutionContext, ExecutionContextExecutor, Future }
@@ -33,24 +34,6 @@ class V2Spec extends BaseSpec with ForAllTestContainer {
   override val container: PostgreSQLContainer = PostgreSQLContainer
     .Def(dockerImageName = DockerImageName.parse("postgres"), urlParams = Map("currentSchema" -> cosSchema))
     .createContainer()
-
-  /**
-   * create connection to the container db for test statements
-   */
-  def getConnection: Connection = {
-    // load the driver
-    Class.forName("org.postgresql.Driver")
-
-    DriverManager.getConnection(container.jdbcUrl, container.username, container.password)
-  }
-
-  // helper to drop the schema
-  def recreateSchema(): Unit = {
-    val statement = getConnection.createStatement()
-    statement.addBatch(s"drop schema if exists $cosSchema cascade")
-    statement.addBatch(s"create schema $cosSchema")
-    statement.executeBatch()
-  }
 
   lazy val config: Config = ConfigFactory
     .parseResources("migration.conf")
@@ -94,7 +77,7 @@ class V2Spec extends BaseSpec with ForAllTestContainer {
   }
 
   override def beforeEach(): Unit = {
-    recreateSchema()
+    recreateSchema(container, cosSchema)
   }
 
   override protected def afterAll(): Unit = {
